rename benchmark package
