+++ bench/server/entry.go
@@ -12,31 +12,33 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package client
+package server
 
 import (
 	"errors"
 
-	"github.com/pairmesh/pairmesh/benchmark"
-	"github.com/pairmesh/pairmesh/benchmark/config"
+	"github.com/pairmesh/pairmesh/bench"
+	"github.com/pairmesh/pairmesh/bench/config"
 )
 
-// Job is the client interface that could be started with Start function
+// Job is the interface of a server, which could be started with Start function
 type Job interface {
 	Start() error
 }
 
-// Run function connects to endpoint host and start performance testing
-func Run(cfg *config.ClientConfig) error {
+// Run function starts a simple echo server as server side of pairbench
+// If cfg.isBounce is true, it echoes whatever it hears from incoming connection
+// Otherwise, it always echoes "OK" as minimal backward payload
+func Run(cfg *config.ServerConfig) error {
 	var job Job
 	switch {
-	case cfg.Mode() == benchmark.ModeTypeEcho:
-		job = NewEchoClient(cfg)
+	case cfg.Mode() == bench.ModeTypeEcho:
+		job = NewEchoServer(cfg)
 		return job.Start()
-	case cfg.Mode() == benchmark.ModeTypeRelay:
-		job = NewRelayClient(cfg)
+	case cfg.Mode() == bench.ModeTypeRelay:
+		job = NewRelayServer(cfg)
 		return job.Start()
 	default:
-		return errors.New("invalid mode specified when starting client (supported mode: echo/relay)")
+		return errors.New("invalid mode specified when starting a server (supported mode: echo/relay)")
 	}
 }
