+++ bench/server/relay.go
@@ -17,13 +17,12 @@ package server
 import (
 	"context"
 	"crypto/rsa"
-
 	"fmt"
 	"time"
 
 	"github.com/flynn/noise"
-	"github.com/pairmesh/pairmesh/benchmark/config"
-	"github.com/pairmesh/pairmesh/benchmark/utils"
+	"github.com/pairmesh/pairmesh/bench/config"
+	"github.com/pairmesh/pairmesh/bench/utils"
 	"github.com/pairmesh/pairmesh/internal/relay"
 	"github.com/pairmesh/pairmesh/message"
 	"go.uber.org/zap"
