+++ bench/utils/rand.go
